Fix factory wipe on first flash
TWRP was using the existance of the RECOVERY_SDCARD_ON_DATA variable as true instead of looking at the value.  So
the data partition was never getting properly reformatted on a wipe.

Another problem was the metadata partition.  For some reason on android make_ext4 assumes block 1024 for the
journal on small filesystems and if it's not passed on the command line.  TWRP doesn't pass the paramater on the
command line so ext4 filesystems smaller than 8MB can't be built without specifying the journal block on the
command line.  I switched /metadata to using ext2.  I doubt anyone is using /metadata at all, but we can revisit
this later if someone cares.